PY-44858 Move placeFile to RefactoringUtil
