+++ .idea/vcs.xml
@@ -51,6 +51,5 @@
   </component>
   <component name="VcsDirectoryMappings">
     <mapping directory="" vcs="Git" />
-    <mapping directory="$PROJECT_DIR$/android" vcs="Git" />
   </component>
 </project>
